requirements.txt
# ============================================================
# FILE: requirements.txt
# PURPOSE: All Python dependencies for the Fintel project.
# DEPENDS: Python 3.11+
# NOTE:    sqlite3 is part of the Python standard library —
#          no pip install needed.
# ============================================================

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
pdfplumber>=0.11.0

# LLM — OpenAI gpt-4o
openai==1.35.0

# Backend
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Frontend
streamlit==1.35.0

# Utilities
python-dotenv==1.0.1
